Extract out the stream_json function
From the export_data module.
